[wlan][policy] Remove connection selection sender error log
The connection selector logs an error when it can't send its selection
to the receiver. This is expected to occur when connection selections
are cancelled and the receiving end of the channel is dropped.

This change removes that error log, as the connection selector should
not care if the receiver is dropped. True errors passed through the
channel will be logged by the receiver.

Bug: 369662954
Test: wlancfg-tests
Change-Id: I944b6c710d50c34848b978658b5d7c8151c16c21
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1125768
Reviewed-by: Marc Khouri <mnck@google.com>
Fuchsia-Auto-Submit: Hayden Nix <haydennix@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>