[dl] Rename TraverseDeps to ReifyModuleTree
TraverseDeps is renamed into ReifyModuleTree, which will check if the
module's module_tree() is set before calling the enqueuing logic to
construct it.

The LinkingSession will use the root_module's module_tree() for the
subset of local-scope modules needed for relocation.

The RuntimeDynamicLinker will call ReifyModuleTree before returning a
pointer to module, if the module was already loaded.

Tests are enabled that require a loaded dependency to have a constructed
module tree when it is dlopen-ed.

Bug: 354786114
Change-Id: I45a325ff4971602101d381b41c6bc0b43e524b4a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1124681
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>