[netstack3] Use weak host semantics when a router
By default, NS3 acts as a strong host. However with this change, it
begins acting as a weak host when forwarding is enabled (i.e. when
configured as a router).

The newly permitted weak-host behavior is:
  * When sending packets, allow a src addr that is not assigned to the
    output device, provided it is assigned to another device that has
    forwarding enabled.
  * When receiving packets, allow a dst addr that is not assigned to the
    input device, provided it is assigned to another device and the
    input device has forwarding enabled.

In both cases, the packet passes through the forwarding hook of the
filtering engine, because conceptually we're "internally forwarding" the
packet from device A to device B. Note that this CL also updates the
integration tests to prove this.

Note, the send path requires checking if forwarding is enabled on a
device during route lookup. To accommodate this, I adjusted the lock
ordering so that the device configuration can be locked after the
routing table has already been locked.

Multiply: netstack-forwarding-integration-test

Bug: 368343873

Change-Id: Id055b44ee682eab8d790129f27facd49d6bfeb72
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115635
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>