[netstack3] Propagate controller removal to Core
Currently, when a fuchsia.net.filter/NamespaceController channel is
closed, the filtering resources owned by the controller are removed in
Bindings, and events indicating the removal are dispatched to watching
clients, but the removal is not actually committed to Core. The state in
Bindings and Core remains out of sync until the next commit to the
filtering state, at which point Bindings's updated state is committed to
Core in its entirety, resolving the discrepancy.

Instead, resources should be removed from Core immediately when a
controller is dropped. Flesh out the existing test that exercises the
controller lifetime semantics to actually verify the side effects of
filter installation and removal, by attempting to connect to a TCP
socket listening on a local port, both before and after the controller
is dropped (when the controller is alive, the connection attempt is
blocked, and when it's dropped, the filter is removed and the connection
attempt succeeds).

Fixed: 369836595
Test: netstack-filter-fidl-tests
Change-Id: Idffe8437c0af01cdcb600ac2aa6387b56b9a463e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126720
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>