[owners] Remove glamrock@google.com from src/diagnostics/config/fire/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/glamrock@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I28556d5d298f8f5e47b71d8913779d13645254b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1118447
Reviewed-by: Ina Huh <ihuh@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Danny Rosen <dannyrosen@google.com>