[f2fs] Improve read performance
This CL tracks recently used pages, and excludes
them from readahead I/Os on read requests. Also,
it creates a VMO as large as needed to supply
pages to the area of a pager's VMO where pages
are not likely to be present. This way
effectively avoids unnecessary readahead I/Os
and the oversupply of VMOs to the area of
a pager's VMO to which it has already supplied
pages before. It uses the hint of the recently
used pages only when memory pressure is low.
On high memory pressure, it assumes that all
pages are not present on the read request area
of a pager's VMO, and does not readahead. Also,
it clears clean pages on FileCache of
pager-backed vnodes during writeback to speed
up lookup for dirty pages on fsync().

Fixed: b/368139435

Change-Id: I5f3f1abd478221694246b97fb43b0da6fb696453
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116913
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Seokhwan Kim <sukka.kim@samsung.com>