[owners] Remove kevinwells@google.com from src/sys/lib/fuchsia_backoff/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/kevinwells@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I9be90354ea400b744054ff144a6bc56c485e087d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1127729
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Danny Rosen <dannyrosen@google.com>