[power] level and dep tracking for blame analysis
This patch to the prototype tool introduces lifecycle
tracking for elements and dependencies to enable blame
analysis on a per-event basis.

It then identifies a minimal-energy duration in each
suspend/resume window and runs a flow-analysis from active
levels of execution_state.

Any element level that contributes to an non-zero level of
execution_state is reported. As events progress
within a minimal-energy duration, the user should observe
a point where the system failed to reduce levels further.

This doesn't quite get us to "print leases" as promised in
the bug, but it's quite close.

Since we have a finite window of captured events, the
processing must be robust against missing events before
history and after history. We handle such cases as we
encounter them.

Bug: 357909949

Change-Id: I9b6fdc0311b897b0fda66ca75c9c000de44885da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122221
Reviewed-by: Prashanth Swaminathan <prashanthsw@google.com>
Commit-Queue: Jaeheon Yi <jaeheon@google.com>