[netcfg] Remove input interface from masquerade API
Currently the fuchsia.net.masquerade/ControlConfig struct includes an
input interface, documented as "the interface carrying the network to be
masqueraded". However, in practice, this interface does not have an
effect on the implementation: the masquerade NAT functionality provided
by the netstack only allows matching on an egress interface, because the
ingress interface is not available in the POSTROUTING hook, where
masquerade NAT is performed.

The input interface was used to call
fuchsia.net.filter.deprecated/Filter.EnableInterface, but enabling
filtering on the input interface is neither necessary nor sufficient for
masquerading to work, as described above. Currently NetCfg *also*
enables the output interface, and that *is* load-bearing for
masquerading to be correctly performed.

To clarify the semantics of the API, remove the input interface field
from the fuchsia.net.masquerade API entirely and only call
Filter.EnableInterface on the output interface.

Test: policy-management-integration-test
Change-Id: I51ad64744f159df639e1257556b8e019e1aad5c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126674
Reviewed-by: Zoey Bush <zmbush@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>