[wlan][policy] removed unused past connection data
The `connect_attempt_time` and `time_to_connect` values that are saved
as part of `PastConnectionData` are currently unused. This change
removes them, to simplify some of the multi-bss support in state
machine. They can be added back in the future if there is a use for
them.

Bug: None
Test: fx test wlancfg-tests
Change-Id: Idb6b637f9a90eec7e5816dddf26108055cef663f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1123883
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Naomi Mccracken <nmccracken@google.com>
Fuchsia-Auto-Submit: Hayden Nix <haydennix@google.com>