[zircon][docs] Remove mention of ZX_CPRNG_DRAW_MAX_LEN.
This constant is an implementation detail of the vdso and syscall
interface, and should not significantly affect the behavior of callers.
There are a few instances where the name seems to have suggested to
callers that they can't pass a buffer larger than this value, so
we should just not mention it to avoid confusion.

Bug: 369851484
Change-Id: Ib2756be9e4e2dc7efc6bf1bf41e587fd92c1f3f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126632
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>