[owners] Remove glamrock@google.com from src/diagnostics/config/persistence/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/glamrock@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ib01b346323fd57ad8a793546f9c8dc01dcd516df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1118542
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Danny Rosen <dannyrosen@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Danny Rosen <dannyrosen@google.com>