[netstack3] Validate netstack_thread_count usage
The netstack_thread_count setting is only meaningful for Netstack3, but
it's currently possible to specify it for Netstack2. Since that doesn't
have any effect, it's almost certainly unintentional and we want to
prevent it from happening.

This CL just returns an error in this case (or, more specifically, when
Netstack2 is selected and not overridden by a Netstack3-only board). I
tried an alternative approach where the netstack_thread_count option was
placed under NetstackVersion, but this led to two primary issues:

1. It made it unintuitive to change the netstack version. In GN,
   Netstack2 could be specified with:

       netstack_version = "netstack2"

   but Netstack3 would require:

       netstack_version = netstack3 = {}

   or the equivalent JSON.

2. Because the above is a breaking change, it would require updating
   every place that netstack3 is currently specified (many of which are
   out-of-tree).

Given that the number of times netstack_thread_count is specified should
be approximately 0 (though currently 2), and Netstack2 will be going
away, it seemed like the pain of the more structured solution wasn't
worth it.

Fixed: 360393264
Change-Id: I76ff67655279443da86f10ec8539cf5a8d2b3dc7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128753
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>