[pkg-cache] Clear variant when loading cache packages
We already do this for base packages, but not cache packages.
The full package url of cache packages are only used for inspect and
to serve the cache package index, where it clears the variant on the
spot, so it needed to clone the map.

Change-Id: I2f66e80bbaeaa49a1f90db2df5acec86d12cbda0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126056
Reviewed-by: Thomas Zander <thomasalpinus@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Sen Jiang <senj@google.com>