[fidl][rust] Decouple resources from core library
This allows the same FIDL bindings to be used on host and target, even
when the host side does not support handles. This also lays the
foundations for other resource types in FIDL, such as driver handles and
FDomain handles.

With this change, it is possible to use the core FIDL library and
bindings without pulling in any Fuchsia-specific dependencies.

Change-Id: Iadf25145ec75847de288185972be0c19b59e1061
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1125893
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: David Koloski <dkoloski@google.com>