[zx][rs] fuchsia-zircon -> zx
This shorter name matches the common convention of
`use fuchsia_zircon as zx;` that is used so commonly in the
tree. Also aligns with the name of the C++ syscall wrapper
library libzx.

Also moves the library to a more suitable location for
eventual inclusion in a Rust SDK.

Leaves forwarding targets with the old names so that
the tree can be migrated incrementally.

Change-Id: I5ff01cbd6c94e35f824a002c608a4f48b8f508e5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128283
Owners-Override: Adam Barth <abarth@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>