[owners] Remove glamrock@google.com from src/security/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/glamrock@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I3440f43702a2f8904549ceff4e6dfb73735f49e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1118543
Reviewed-by: Danielle Kay <danikay@google.com>
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
Commit-Queue: Danny Rosen <dannyrosen@google.com>