[package-tool] Avoid overwriting TUF repo data inside a product bundle
This will copy the TUF metadate into a user-provided empty directory
which serves as the (mutable) TUF metadata store for the blobs in the
product bundle. This ensures the original TUF metadata in the product
bundle is not modified when the blobs are served via the repo server.

Bug: b/335008631
Change-Id: Ib90e5a8a14d7bb14aadf192c3a8c790a367bb0af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128015
Commit-Queue: Thomas Zander <thomasalpinus@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>