[sdmmc] Clean up GN targets
Create a "source_set" GN target containing the source files and
dependencies of the sdmmc driver. Make the GN targets of the sdmmc
driver binary and sdmmc tests depend on this new target.

This makes it easier to ensure that the sdmmc driver binary and tests
have the correct dependencies.

Change-Id: I9b92f200e733c30df23229912295e7bca81c931d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1101035
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>