[testrunner] Allow /usr/share/ca-certificates in sandbox.
This is the directory where Debian-based systems store the
contents of the ca-certificates package which has the majority
of recent roots of trust for SSL connections.

For example, this is necessary to allow our symbolizer to
connect to our symbol servers during tests that want to
symbolize backtraces.

Change-Id: Ie06d87a8bb7a6cd7d226bcb78f4653c962fbd8fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129752
Reviewed-by: Ina Huh <ihuh@google.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>