[runtests-utils] Adjust test for profile data
DiscoverAndRunTestsWithOutput test have certain checks for
summary.json file in the generated test data that is only
valid in an uninstrumented build. When we run this test
with instrumentation enabled, we have "data_sinks" section
in the summary.json. This patch adjusts the test, so it
passes in an instrumented build like in profile builders.

Bug: 348396735
Change-Id: I6c05f49e9f5893ccc711e59e3d967a7ad93a7a76
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128281
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Gulfem Savrun Yeniceri <gulfem@google.com>