[ffx] remove BuildOverride
Based on code inspection and some testing, BuildOverride doesn't seem
to do anything for queries. (In particular, regardless of the setting
of the query's "build" parameter, the same config is returned.)  This
suggests that the functionality is not being used, and since it adds a
lot of complexity to the ffx config code, it should be removed.

Another use of BuildOverride is used to implement the setting of
build files in Environment. But in this CL it is simplified down to
an Option<Path>.

Bug: 342038349
Change-Id: I121cdc10bc270a32d336c70d319b161e92e57fa3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126673
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>