[spmi] Switch to debug service in spmi-ctl and update routing
The --target option is now mandatory and accepts an integer ID of the
SPMI target device to connect to. An optional --controller option is
added to specify the name of the controller to connect to: if
unspecified, the first entry in the service directory is used. The
ability to connect using a hexadecimal service name has been removed.

All routes that previously referred to TargetService and
SubTargetService have been changed to DebugService to reflect the
changes to spmi-ctl.

Test: spmi-ctl-test
Test: Used spmi-ctl to read registers on Kola
Fixed: 368117439
Bug: 361853516
Change-Id: I4de90d86684a973c3a8d971e1e64647db64e6376
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128274
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Fuchsia-Auto-Submit: Braden Kell <bradenkell@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>