[logs] Use random per-boot integer for detecting reboots instead of UTC
The UTC to monotonic offset can change frequently and is
not guaranteed to stay the same within the same boot.
Instead of using this to detect reboots, use a per-
boot random number that is generated from the monotonic
clock when Overnet starts on the device.

Fixed: b/370548387
Change-Id: I343cfff6c988ad45b60fc92625556d17c093c2c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128476
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>
API-Review: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>