[dl] Improve test logic for BasicGlobalDynamicTls
The previous version of the string matching only matched the glibc
format. The only reason this worked is because these tests were only
running on the host, and were not yet using the musl libc, since it had
a bug in its handling of undefined weak symbols.

Additionally, the host toolchains are able to select the TLS dialect for
either traditional TLS or TLSDESC. Because of those factors, we only
ever saw the glibc version of this error, until we tried to enable these
tests for Fuchsia targets in https://fxrev.dev/1111133, and had bots
running in CI that were expected to pass. What still isn't clear is why
some arm64 bots were able to pass this test (or why they skipped it?).

This patch updates the matching for the error string when
get_tls_dep_data is not available. Additionally, it incorporates related
test clean ups to the logic of the HAVE_TLSDESC block, which was always
silently skipping that code. Finally, it groups some of the checks more
logically.

Bug: 338295240, 325518247, 342480690
Change-Id: I245a82dcb832fcb4e7edf763fbb86e140c67a959
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1118853
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Paul Kirth <paulkirth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>