[starnix] Add test for capabilities.
This CL adds a new test for the capabilities system calls.

The test has to test how certain capabilities properties behave across
calls to `execve`, so the CL also adds a new binary named
"print_helper". This binary can be invoked with different parameters to
print information of the child process. The output of this
`print_helper` is parsed by the capabilities test.

A bunch of functions to analyze and modify the capability sets have been
added to `capabilities_helper`, now part of `test_helper`.

Change-Id: I148f25934b15c22e96b1b744bf745815f8aa421b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1120193
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>