[netstack3] Perform correct type of NAT in all hooks
Currently, only Redirect NAT (a special case of DNAT) is supported by
Netstack3. DNAT is only configured in the LOCAL_EGRESS and INGRESS
hooks, which is important because rewriting the destination of a packet
affects routing, so it must be done before a final routing decision is
made.

However, for reply packets on a connection that is DNATed, their source
address/port is rewritten to match the original destination of the
connection. This source rewriting must occur in LOCAL_INGRESS or EGRESS,
i.e. the same hooks where SNAT can be configured. Today, NAT rewriting
of packets in both the original and reply directions of a connection is
performed in the DNAT-specific hooks, which is incorrect (and can cause
more serious issues for, e.g., Masquerade NAT).

Change the filtering logic such that NAT is *performed* in every hook,
but can still only be *configured* for new connections in the
DNAT-specific hooks (LOCAL_EGRESS and INGRESS).

Bug: 341771631
Test: netstack3-filter-test
      netstack-filter-fidl-tests
Change-Id: Ifb1fd26f88966bb4704182566e83efdfcaddb934
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129760
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Peter DeLong <pdelong@google.com>