[build][fidl] Limit unstable to internal and partner categories
Only allow FIDL libraries to be unstable when the SDK category is
"internal", "partner", or not specified. This specifically disallows
unstable libraries in the "partner_internal" category, which as noted in
https://fxbug.dev/364727245, does not make sense.

Move `fuchsia.wm` to the "internal" SDK category to meet this condition.

Allow exceptions for known exceptions in "partner_internal" due to ffx
build requirements.

Bug: 325669391, 364727245
Change-Id: I15209b0e7cba5cad741260f25f3c5e762bd88540
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1125780
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>