[sdk][devices] Adding count field to cpu definition
This adds a "count" field to the cpu definition in the
hardware schema. This value is used by the emulator to
configure the number of CPUs to define when running the
Fuchsia instance.

Previously, it was hardcoded to 4, so as a backwards compatibility
method, the cpu.count field defaults to 4 if it is not present when
deserializing the data.

Bug: 369114507
Change-Id: I6914c96a1598b3940aa3bc92e5f24ac736b29f33
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1124052
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>