[ffx][repo] Fixing product-bundle server registration
This fixes 2 issues related to package serving with a product
bundle as the repo source.

First, it constructs a repository name based on the "base repo name",
which is either the `--repository` option, the default configured repo,
or devhost. Then appends `.` and then first alias for each repository
configured in the product bundle. This avoids having multiple servers
named "fuchsia.com" and serving different packages.

Second, this adds registering the aliases configured with the repository
when registering the repo with the device. Previously, since the name of
the repo was the first alias, it would work for 1 product bundle with
repositories that had 1 alias each.  Now it works for multiple product
bundles.

Bug: 370594742
Bug: 370089652
Change-Id: Id419045db1e08abed4cdd36d1bec0f5146fc0043
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128280
Reviewed-by: Thomas Zander <thomasalpinus@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>