Reland "[aml-mipi] Use sdk pdev library"
This is a reland of commit bb50c4a5db7e563653864f50535d2c9961bb5f8e

The original CL made the aml-mipi driver try to connect to a device
fragment called "pdev". The aml-mipi is not a composite driver and does
not have device fragments. This CL updates aml-mipi driver to try to
connect to its parent device instead of a fragment device. This CL also
adds tests for the aml-mipi driver to verify that the driver can
successfully bind.

Original change's description:
> [aml-mipi] Use sdk pdev library
>
> Make the aml-mipi driver use the //sdk/lib/driver/platform-device/cpp
> library instead of //src/devices/bus/lib/device-protocol-pdev.
>
> Bug: 366517646
> Change-Id: I89d872ff04d69085eabdc491b22c600b50357a0d
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1118693
> Reviewed-by: Novin Changizi <novinc@google.com>
> Commit-Queue: Nick Eaton <nickeaton@google.com>

Bug: 366517646
Change-Id: I7dda8e476714651d0ef6aacbf26652782a51f66f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122752
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>