[netemul] Disambiguate protocol names using source
Currently, netemul installs all capabilities provided a realm's children
to the root node. This introduces panics when multiple components expose
the same protocol.

This change adds a suffix to each exposed protocol to ensure that
clients can specify the name of the protocol they would like to connect
to, or leave it unspecified to connect to the first component offering
that protocol.

Adds a `counter-alternative` component for testing to demonstrate
which protocol from which child will be used when they both offer the
same protocol. One component has program_args specified for the starting
value so that we can determine which one is being used.

There will be a follow-up CL to allow specifying child name in the
client netemul API.

Bug: 42151461

Change-Id: If3f6fe0aa9bcfe651c168c5e560155b05a1eeee1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126658
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>