[netemul] Add endpoint for protocol child name
Users of the netemul TestRealm API can specify a particular child to use
when connecting to a protocol.

Add a test to ensure that specifying the child component's name when
connecting to a protocol results in creating a proxy that returns the
expected results.

Fixed: 42151461
Change-Id: I52bfa9f8ac17cadce7784b9b15fb06fb97d5ab07
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129975
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>