[bt][fidl] Add generic UUID to BR/EDR service search params
Currently, the BR/EDR Profile Search parameters require a mandatory
profile/service identifier. This identifier is actually a UUID and the
Profile FIDL API provides a list of known IDs for profiles & services. In
some cases, it is desirable to search for a custom UUID instead.

Add an optional field, `full_uuid` to the service search parameters. If
specified, the `bredr.Profile` server will use this UUID as the search
filter. It is invalid to specify both a `full_uuid` and `service_uuid`
in the parameters - an error will be returned in this case.

Bug: 355487849
Test: API only - fx build

Change-Id: I9ddd20e67f993d20a62a60b2f80f475232a68790
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110654
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
API-Review: Marie Janssen <jamuraa@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>