[kernel][crashlog] Fix reporting a disabled debuglog.
When a crashlog is generated, and the kernel requests that the
debuglog be added to the crashlog, whether or not that actually
happens is gated on the `render_dlog_to_crashlog` boot option.  When
the option was not set, the crashlog was _supposed_ to contain a clear
indication that the log was requested, but deliberately not rendered
because of the boot option.

This was not working correctly; fix that issue.

Bug: 364882116
Change-Id: I24c146e41275b6840f9ed6d82de2001466a3480f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130772
Reviewed-by: Francois Rousseau <frousseau@google.com>
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>