[driver_runtime][rust] Add driver server
Implements a `DriverServer` that interfaces between the driver host and
the rust driver object, including managing its async lifetime. This is
analogous to the C++ `DriverServer`.

Also adds the basic outline of a `Driver` trait that is kind of like
C++'s `DriverBase`, except that instead of holding the context in a base
class it is given a `DriverContext` with that information.

As part of this, this also breaks up the main `fdf_rust` library so that
`fdf_sys` is its own (visibility-gated) library that both `fdf_rust` and
`fdf_rust_server` depend on. This avoids a dependency loop since the
server depends on generated driver transport fidl, which in turn depends
on `fdf_rust` for things like Channels.

Fixed: 359228261
Change-Id: I06edf730324d06320b44dd2693502d1f1ce0f3ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1090314
Reviewed-by: David Koloski <dkoloski@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>