[ffx] `ffx target list` should report port
When the port is not the default target ssh port (22), ffx target list
should display the port. Among other things, this will allow
programs that wish to provide the IP:port as the target (which
is a requirement for `ffx-strict` invocations) to use discovery
to find the appropriate port.

Bug: 42154593
Change-Id: Iad47916f537cccb9e9717943136f830084626431
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129908
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>