[net-cli] Add if config subcommand
Add an `if config` subcommand for getting/setting interface
configuration. The getter returns all interface configuration params
and their values. The setter takes pairs of a parameter name and the
value to be set and applies the configuration.

Make `if config set` support setting the config knob for disabling
SLAAC temporary address generation.

The existing subcommands `igmp`, `mld`, and `ip-forward` are henceforth
considered deprecated and will soon be implemented in terms of
`if config` instead.

Change-Id: I93353209b2690ec1c5242753075bf6e33905d777
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108835
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Tony Gong <gongt@google.com>