[ffx][fho] Exposing host_ssh_address for direct connection
This adds a method to get the host_ssh_address from the overnet
connection when running directly, without the daemon. This is the
address of the host from the device's perspective and is used to
address resources on the host, such as package servers.

Change-Id: I3ffb1899a3572f9c9109b195b8f33bb7f05a9731
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130020
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>