[ffx][emu][crosvm] Add limited crosvm engine backend support
Support is highly limited. You can only boot with the serial console
being routed to stdio and no audio, graphics, input, etc devices are
supported. Future changes may improve support. For now, support is
sufficient to boot the microfuchsia and minimal products.

To minimize the amount of effort required, it was implemented as if it
were qemu based as there exists a lot of code in the qemu_based module
which is helpful even for crosvm. Major refactoring is required to
decouple that code from qemu_based.

This change also adds the crosvm tool to the sdk.

Test: ffx emu start --engine crosvm -H --console --net tap
  ffx target ssh
  ffx emu stop
Test: ffx -c product.path=$MICROFUCHSIA_PB emu start -H --console
  socat TCP4-LISTEN:52925,reuseaddr,fork VSOCK-CONNECT:3:22
  ffx target add 127.0.0.1:52925
  ffx component show
Test: fx test ffx_emulator_engines_lib_test
Bug: 348473396
Change-Id: I560e9974b514cb72036f1117d24de66d6e0abdfa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129861
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>