[owners] Remove kevinwells@google.com from src/lib/testing/fuchsia-hyper-test-support/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/kevinwells@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Iaab1f79c26229a932f0d36c5b95cd2067f949b36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1127728
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Danny Rosen <dannyrosen@google.com>
Reviewed-by: Danny Rosen <dannyrosen@google.com>