[gpt-rust] Restore primary metadata from backup
This is required by the UEFI spec (see 5.3.2).  It also ensures that we
are resilient to sequences of partially completed updates; for example,
if the following sequence happens, primary restore is necessary to
ensure the GPT remains consistent:

1. Metadata is updated.  Only the backup metadata is fully written, and
   the primary metadata is partially written.
2. The system boots and uses the backup metadata.
3. (This is when primary ought to be restored.)
4. Metadata is updated, but this time, the backup metadata is partially
   written.

Without restoring the primary metadata in step (3), the GPT would be
corrupt after this sequence, as both copies would be corrupt.

Bug: 371060853
Change-Id: Ia3399fe48f0233e4576549f99447ef95c3167aed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1131492
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>