[ffx] Fix default target test
fxr/1115615 changed the environment to always set FUCHSIA_NODENAME.
It turns out that this breaks a test that was making some assumptions
about the environment. This change fixes the test to exlicitly clear
out the relevant environment.

(The code in question also looks at EnvironmentContext for env vars, so
this CL changes the context as well.)

Fixed: 372035158
Change-Id: Ibe31f9ee405b1818ef2ce2e82541c79a270cd830
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1132356
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>