[driver_loader] Move bootstrap logic out of driver manager.
The driver manager was writing some stuff to the bootstrap
channel for the test driver host to read, but we don't
want to do the same thing in the real driver host.

The loader will handle creating the bootstrap channel pair
and in future sending the required bootstrap messages.

The loader constructor also takes in a handler that is
called after every driver is loaded. This allows tests to
inject their customized messages to communicate with the
fake driver host test binary implementation.

Multiply: driver-host-runner-tests :50

Bug: 341997294

Change-Id: Ic9afc15c76043e377fb9c8db564981c820ef406f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128320
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>