[remote-control] Add default identifier
When used in embeddable configurations, rcs does not have access to most
of the protocols it uses to determine identity. As a result we need an
alternative approach to determining identity. For now, we can use a
hardcoded identifier. This change in the future once we have better ways
to determine uniquely determine identity of the system.

Bug: 335549697
Change-Id: Ib1f0efb0b283a9a00e2771cef87c293e0942687f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128279
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>