[functionfs][adb] Add Proxy wake support
Add support for proxying the ADB channel so that new
information will wake the starnix container.

The actual proxying is disabled in this CL and will
be enabled later.

This also fixes an existing issue in the code where
events from the ADB stack were never read. They need
to be read once wake proxying is enabled. The code
spawns a thread which reads and discards these messages
in a loop.

Bug: b/369114941
Change-Id: I3e6fbfaef54f40ce754b90ffaeaad5cd3fe9cd7c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129772
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>