[components][sandbox] Support DirEntry on host
Support on host to prepare for typed routers, including DirEntryRouter.

The host version of this type is basically empty, but it allows code
that uses this type to compile on host. In practice, the host code will
use debug routing so it doesn't need DirEntry to be functional.

Bug: 360221168

Change-Id: I25c11a2635f8f3f7485e006f4f45858af07afc63
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1107927
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>