[kernel][crashlog] Print boot time in crashlog
As dictated by RFC-0260, use the current_boot_time() to indicate system
uptime in the crashlog. However, the user may also want to know how long
the system was up not including periods of suspension, so add a line
that prints current_time() as well; this quantity is now called
"RUNTIME" to reflect the fact that it is the amount of time the system
was awake and running.

This change does not add RUNTIME to the RAM mappable crashlog; that is
left to a future change.

Bug: 328306129
Change-Id: Ie31dfa6ca0975a42228a166245a62e467bbd5e42
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130234
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>