[tee] Stub TEE_UnmaskCancellation
Instead of panicking, return true from TEE_UnmaskCancellation. This is
needed to pass preconditions for some tests, even though the tests
themselves do not rely on this functionality. Proper implementation of
Cancellation APIs is left as a TODO.

As a result of this change, regression1038 no longer panics and fails.
It talks to the TA successfully and determines to skip the test because
it doesn't support MTE.

Bug: b/361406102
Test: fx test //src/tee
Change-Id: I243c8d0bf8259891008a490e917df9da86c827c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1132523
Fuchsia-Auto-Submit: Eric Stone <ecstone@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Eric Stone <ecstone@google.com>